service/wip -> refactor transport/rest, delete middleware, moved the logic to the handler
