+++ internal/transport/rest/auth.go
@@ -6,14 +6,12 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/zenorachi/image-box/internal/service"
 	"github.com/zenorachi/image-box/internal/transport/logger"
-	"github.com/zenorachi/image-box/internal/transport/rest/middleware"
 	"github.com/zenorachi/image-box/models"
 	"net/http"
-	"strings"
 )
 
 func (h *handler) signUp(ctx *gin.Context) {
-	inputBodySignUp, _ := ctx.Get(middleware.InputSignUp)
+	inputBodySignUp, _ := ctx.Get(inputSignUp)
 	input, _ := inputBodySignUp.(models.SignUpInput)
 
 	if err := input.Validate(); err != nil {
@@ -32,7 +30,7 @@ func (h *handler) signUp(ctx *gin.Context) {
 }
 
 func (h *handler) signIn(ctx *gin.Context) {
-	inputBodySignIn, _ := ctx.Get(middleware.InputSignIn)
+	inputBodySignIn, _ := ctx.Get(inputSignIn)
 	input, _ := inputBodySignIn.(models.SignInInput)
 
 	if err := input.Validate(); err != nil {
@@ -73,42 +71,4 @@ func (h *handler) refresh(ctx *gin.Context) {
 
 	ctx.Header("Set-Cookie", fmt.Sprintf("refresh-token=%s; HttpOnly", refreshToken))
 	ctx.JSON(http.StatusOK, gin.H{"token": accessToken})
-}
-
-func (h *handler) CheckToken() gin.HandlerFunc {
-	return func(ctx *gin.Context) {
-		token, err := getTokenFromRequest(ctx)
-		if err != nil {
-			logger.LogError(logger.AuthMiddleware, err)
-			ctx.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "JWT not found"})
-			return
-		}
-
-		id, err := h.userService.ParseToken(ctx, token)
-		if err != nil {
-			logger.LogError(logger.AuthMiddleware, err)
-			ctx.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{"error": "cannot parse token"})
-		}
-
-		ctx.Set("userID", id)
-		ctx.Next()
-	}
-}
-
-func getTokenFromRequest(ctx *gin.Context) (string, error) {
-	header := ctx.Request.Header.Get("Authorization")
-	if header == "" {
-		return "", errors.New("empty authorization header")
-	}
-
-	headerParts := strings.Split(header, " ")
-	if len(headerParts) != 2 || headerParts[0] != "Bearer" {
-		return "", errors.New("invalid authorization header")
-	}
-
-	if len(headerParts[1]) == 0 {
-		return "", errors.New("token is empty")
-	}
-
-	return headerParts[1], nil
 }
